refactored AddAllocRound
